# lib/api-zod/src/generated/types/authUserResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Task Manager Suite API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthUserResponse {
  user: User;
  token: string;
}