# lib/api-zod/src/generated/types/clearCompletedResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Task Manager Suite API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClearCompletedResult {
  deleted: number;
}