# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Task Manager Suite API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputPriority } from './taskInputPriority';
import type { TaskInputStatus } from './taskInputStatus';

export interface TaskInput {
  title: string;
  /** @nullable */
  description?: string | null;
  status?: TaskInputStatus;
  priority?: TaskInputPriority;
  /** @nullable */
  dueDate?: string | null;
}