# lib/api-zod/src/generated/types/todoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Task Manager Suite API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TodoInputPriority } from './todoInputPriority';

export interface TodoInput {
  title: string;
  /** @nullable */
  description?: string | null;
  priority?: TodoInputPriority;
  /** @nullable */
  dueDate?: string | null;
}